ver 0.06

+++ src/main/java/Manager/CarService.java
@@ -2,15 +2,16 @@ package Manager;
 
 import DAO.CarDAO;
 import Model.Car;
+import Model.CarRentDetails;
 import Model.CarType;
 import Utils.StringUtils;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class CarService {
 
     CarDAO carDAO = new CarDAO();
-
     static StringUtils strUtil = new StringUtils();
 
     public List<Car> getAllCars() {
@@ -18,9 +19,8 @@ public class CarService {
         return allCars;
     }
 
-    public String addNewCar(Car vehicle) {
+    public String addNewCar(Car vehicle, CarRentDetails details) {
 
-        // sprawdzenie czy carPlateNumber nie ma !^*($@
         String plateNumber = strUtil.checkSring(vehicle.getCarPlateNumber());
         vehicle.setCarBrand(vehicle.getCarBrand().toUpperCase());
         vehicle.setCarPlateNumber(plateNumber.toUpperCase());
@@ -29,7 +29,7 @@ public class CarService {
         if (exists) {
             return "SERVICE : Car already exists";
         } else {
-            carDAO.addNewCarToDB(vehicle);
+            carDAO.addNewCarToDB(vehicle, details);
             return "SERVICE : Car saved in DB";
         }
     }
@@ -40,7 +40,13 @@ public class CarService {
         //sprawdzić czy nie zawiera dziwnch znaków @!@$$
         plateNumberToDelete = strUtil.checkSring(plateNumberToDelete);
 
-        carDAO.deleteCarFromDB(plateNumberToDelete);
+        boolean exists = checkIfCarExists(plateNumberToDelete);
+
+        if(exists) {
+            carDAO.deleteCarFromDB(plateNumberToDelete);
+        } else {
+            System.out.println("Sorry, no : " + plateNumberToDelete + " in DB");
+        }
 
     }
 
@@ -56,7 +62,6 @@ public class CarService {
         carType = strUtil.checkSring(carType);
         boolean exists = carTypeExists(carType);
         if (exists) {
-            System.out.println("Tak mamy cos w bazie");
             carDAO.getAllCarsBasedOnType(carType);
         } else {
             System.out.println("NIE mamy takiego carType w bazie");
@@ -64,13 +69,26 @@ public class CarService {
     }
 
     public void getCarsBasedOnBrand(String carBrand) {
+
         carBrand = strUtil.checkSring(carBrand);
-
-
-
         carDAO.getAllCarsBasedOnBrand(carBrand);
     }
 
+    public List<Object> getAllAvaliableModels(String carBrand){
+        List<Object> avaliableList = new ArrayList<>();
+
+        carBrand = strUtil.checkSring(carBrand);
+
+        avaliableList = carDAO.getAllAvaliableModels(carBrand);
+
+        return avaliableList;
+    }
+
+    public List<Object> getAllAvaliableCars() {
+        List<Object> avaliableList = carDAO.getAllAvaliableCars();
+//        avaliableList = carDAO.getAllAvaliableCars();
+        return avaliableList;
+    }
 
     ///////////////////////////////// PRIVATE METHODS /////////////////////////////////
 
@@ -83,6 +101,7 @@ public class CarService {
         }
         return false;
     }
+
 
 }
